Refactors the TreeBinner
